refactor(cliente): padronização para PascalCase e melhorias estruturais
- Renomeadas propriedades das entidades para seguir o padrão PascalCase (ex.: idCidade → IdCidade).
- Adicionadas propriedades na entidade Cliente:
  - EstadoCivilId
  - Ativo
  - Sexo
  - DataNascimento
- Criada a entidade EstadoCivil com relacionamento adequado.
- Criadas as classes ClienteRequest e EnderecoRequest para encapsular dados de requisição.
- Atualizado o mapeamento das entidades para refletir os novos nomes e relacionamentos.
- Adicionados arquivos de migração para:
  - Renomear colunas existentes.
  - Criar novas colunas e tabelas.
- Atualizadas as classes de repositório para utilizar os novos nomes.
- Ajustado relacionamento na entidade StatusEntrega, garantindo integridade referencial.

Esta alteração melhora a padronização, legibilidade e manutenção do código.

+++ RastreamentoPedido.Core/Queries/Clientes/CidadeQueries.cs
@@ -4,7 +4,7 @@ namespace RastreamentoPedidos.Data.Queries.Clientes
 {
     public class CidadeQueries
     {
-        public static QueryParamsSQL obterCidadePorId(int id)
+        public static QueryParamsSQL obterCidadePorId(long id)
         {
             return new QueryParamsSQL
             {
