+++ RastreamentoPedido.Core/Repositories/Cliente/IClienteRepository.cs
@@ -5,7 +5,7 @@ namespace RastreamentoPedido.Core.Repositories.Clientes
     public interface IClienteRepository : IRepository<Cliente>
     {
         Task<IList<Cliente>> CarregarTodos();
-        Task<Cliente> CarregarPorId(long id);
+        Task<Cliente> CarregarPorId(int id);
         Task<Cliente> CarregarPorEmail(string email);
         Task<Cliente> CarregarPorDocumento(string documento);
         Task<Cliente> Adicionar(Cliente cliente);
